dummy_data/requirements.txt
flask
psycopg2-binary
faker
gunicorn